- (enh) user service renamed
